Python/webserver/requirements.txt
Django==1.6.1
South==0.8.4
amqp==1.3.3
anyjson==0.3.3
billiard==3.3.0.13
celery==3.1.7
djangorestframework==2.3.10
kombu==3.0.8
pyserial==2.7
pytz==2013.8
simplejson==3.3.1
wsgiref==0.1.2
